ubuntu/debian: use the same musl installing method
In kata-containers/osbuilder/#412, we invented different ubuntu/debian
dockerfiles for different archs.
Now, we find that a few common improvements introduced to generic file
have been missing in Dockerfile-aarch64.in.
So in order to keep just one dockerfile for multi archs, we'd like to
let ubuntu/debian use the same musl installing method as the other
distros do.

Fixes: #460

Signed-off-by: Penny Zheng <penny.zheng@arm.com>